Perform Routing via ViewModel and Fragment
The nastiness around this at the moment is that we have the same event (launchShowPage) bubbling up from the Interactor, via the presenter, via the view model, to a router that is set up by the fragment because it needs to be context-aware.

Is there a better way to model this event data flow?